log in via fb done

+++ scarlet-api/middleware/CSRFToken.middleware.js
@@ -1,8 +1,12 @@
 const CSRFTokenService = require("../service/CSRFToken.service");
 
 exports.checkCsrfToken = async (req, res, next) => {
+  if (req.url.includes("/users/auth")) {
+    return next();
+  }
   CSRFToken = req.body.CSRFTokenObject;
   try {
+    console.log("URL IN MIDDLEWARE -> -> ", req.url);
     await CSRFTokenService.deleteToken({ CSRFToken });
     next();
   } catch (e) {
